Support generating weights from Rust
- Check out this branch https://github.com/ankitects/anki/pull/2687
- If you want to test with a different version of fsrs-rs, edit Anki's
Cargo.toml and replace the fsrs git with fsrs.path.
- touch Cargo.lock && ./ninja pylib
- Run:

FSRS_RS=1 PYTHONPATH=/path/to/anki/out/pylib:/path/to/anki/pylib python script.py

Note this currently fails to work:
pyo3_runtime.PanicException: ndarray: index 18446744073709551615 is out of bounds in array of len 17

It seems that the generated rating tensor has 0s in it, which causes 0 - 1
to wrap around to a giant index in init_stability(). Maybe I converted the
records incorrectly?