New method of finding the best values of three.py
I came up with a new method of finding the best value of mode, given three values. It reduced the variance of the estimate for both normal and lognormal data, in other words, this estimate is more consistent than the old one, where only two values were used and the third value was discarded.