Changed the order of some algorithms in the table.md (#66)
* Changed the order of some algorithms in the table.md

The problem is that these results are weighted by n(review), but the Wilcoxon test isn't weighted. So the order *can* be different.
My logic is that if the log loss is the same, then such algorithms should be sorted by RMSE.

* Fixed incorrect number of digits of ACT-R (ln(reviews)).md

I don't know why, but the number of significant digits was incorrect. It could be a typo, or it could be that my rounding function has a bug that I missed.

* CI should always have 2 significant digits.py